requirements.txt
adal==1.2.7
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.11.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
argcomplete==2.1.2
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
azure-ai-ml==1.8.0
azure-common==1.1.28
azure-core==1.30.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-data-tables==12.5.0
azure-graphrbac==0.61.1
azure-identity==1.13.0
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.2
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azure-nspkg==3.0.2
azure-openai-tr==0.1.0a5
azure-storage-blob==12.16.0
azure-storage-file-datalake==12.11.0
azure-storage-file-share==12.12.0
azureml-core==1.51.0.post1
azureml-dataprep==4.11.3
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.18.3
azureml-fsspec==1.1.0
azureml-mlflow==1.51.0
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
bs4==0.0.2
build==1.0.3
cachetools==5.3.1
camelot-py==0.11.0
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.23
ci-info==0.3.0
click==8.0.4
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
configobj==5.0.8
configparser==6.0.0
contextlib2==21.6.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.8
Cython==0.29.35
databricks-cli==0.17.7
dataclasses-json==0.5.14
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distlib==0.3.6
distro==1.8.0
docker==6.1.3
dotnetcore2==3.1.23
easyocr==1.7.1
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
ghostscript==0.7
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.20.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.62.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
imageio==2.31.1
img2pdf==0.5.1
importlib-metadata==6.7.0
importlib_resources==6.1.2
ipykernel==6.15.0
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpickle==3.0.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter_core==5.3.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
knack==0.10.1
kubernetes==29.0.0
langchain==0.0.330
langchain-community==0.0.24
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.0.92
lazy_loader==0.2
llama-index==0.10.13.post1
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.5
llama-index-core==0.10.13
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.6
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.6
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.4
llama-parse==0.3.4
llamaindex-py-client==0.1.13
loguru==0.7.2
logzero==1.7.0
looseversion==1.3.0
lxml==4.9.4
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mldesigner==0.1.0b13
mlflow==2.4.1
mlflow-skinny==2.4.1
mltable==1.4.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.5
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.2.0
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numexpr==2.9.0
numpy==1.25.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ocrmypdf==16.0.0
onnxruntime==1.17.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.0.2
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.4
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.10.1
Pillow==10.1.0
pip==23.1.2
pkginfo==1.9.6
platformdirs==3.6.0
pluggy==1.3.0
portalocker==2.7.0
posthog==3.4.2
preshed==3.0.9
prompt-toolkit==3.0.38
protobuf==3.20.3
prov==2.0.0
psutil==5.9.0
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==1.10.14
pydantic-core==2.14.6
pydash==5.1.2
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.7.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pypdf==4.0.2
PyPDF2==3.0.1
pypdfium2==4.27.0
PyPika==0.48.9
pyproject-hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
pyxnat==1.6
PyYAML==6.0.1
pyzmq==25.1.0
querystring-parser==1.2.4
ray==2.0.0
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.3
sentence-transformers==2.5.1
set-loglevel==0.1.2
setuptools==67.8.0
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.16
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.36.3
streamlit==1.31.1
streamlit-chat==0.1.1
strictyaml==1.7.3
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.1
tqdm==4.66.2
traitlets==5.9.0
traits==6.3.2
transformers==4.38.2
triton==2.1.0
typer==0.9.0
typing-extensions==4.10.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
virtualenv==20.23.1
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.6
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.6.0
websockets==12.0
Werkzeug==2.3.6
wheel==0.38.4
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.15.0
